Implemented CoreDocument::generate_method in the bindings
